debug

+++ cyanure/estimators.py
@@ -300,7 +300,7 @@ class ERM(BaseEstimator, ABC):
         if platform.system() == "Windows":
             if scipy.sparse.issparse(X):
                 print("Before conversion : " + str(X.dtype))
-                X.indptr = X.indptr.astype(np.float64).astype(np.uintc)
+                X.indptr = X.indptr.astype(np.float64).astype(np.int64)
                 X.indices = X.indices.astype(np.float64).astype(np.intc)
                 print("After conversion : " + str(X.dtype))
         training_data_fortran = X.T if scipy.sparse.issparse(
